requirement.txt
python==3.6.9
tensorflow-gpu==2.3.1
numpy==1.18.5
